refactor: add new finance and onboarding entities, use cases, and repositories; update imports for quickpay models
