+++ lib/modules/quickpay/data/model/coin_assets.dart
