+++ lib/modules/quickpay/presentation/screens/quick_pay_home_screen.dart
@@ -7,7 +7,7 @@ import 'package:defifundr_mobile/core/routers/routes_constant.dart';
 import 'package:defifundr_mobile/core/utils/resolve_color.dart';
 import 'package:defifundr_mobile/modules/onboarding/presentation/identity_verification/widgets/brand_button.dart';
 import 'package:defifundr_mobile/modules/onboarding/presentation/multi_factor_authentication_screen/widgets/custom_back_button.dart';
-import 'package:defifundr_mobile/modules/quickpay/presentation/class/quick_payments.dart';
+import 'package:defifundr_mobile/modules/quickpay/data/model/quick_payments.dart';
 import 'package:defifundr_mobile/modules/quickpay/presentation/widgets/checkbox_status.dart';
 import 'package:defifundr_mobile/modules/quickpay/presentation/widgets/filled_quickpay.dart';
 import 'package:defifundr_mobile/modules/quickpay/presentation/widgets/filter_buttons.dart';
